Revert "fix(auth,platform): add authorization of rbac in inspectors (#1089)"
This reverts commit 644dec56dc57970d425b6deef94e3b638657ff2d.
